requirements.txt
numpy==1.26.4
numpy-quaternion==2023.0.4
opencv-python==4.10.0.84